refactor: split occurrences.rs into submodules
Break the 739-line occurrences route file into focused submodules:

- read.rs: GET handlers (nearby, feed, bbox, geojson, occurrence, observers)
- write.rs: POST/DELETE handlers (create, delete, observer management)
- auto_id.rs: auto-identification record building logic
- mod.rs: re-exports all public handlers

All existing import paths (routes::occurrences::*) remain unchanged.
Pure reorganization with no logic changes.
