+++ crates/observing-appview/src/routes/occurrences/mod.rs
@@ -1,0 +1,6 @@
+mod auto_id;
+mod read;
+mod write;
+
+pub use read::{get_bbox, get_feed, get_geojson, get_nearby, get_occurrence_or_observers};
+pub use write::{create_occurrence, delete_occurrence_catch_all, post_occurrence_catch_all};
